Refactor away more C# classes.
