+++ scenes/weapons/missile_emitter.tscn
@@ -5,6 +5,6 @@
 
 [node name="MissileEmitter" type="Node2D" groups=["weapon"]]
 script = ExtResource("1_6k2de")
-ProjectileBaseItem = ExtResource("2_ag4ag")
+Projectile = ExtResource("2_ag4ag")
 ImpulseStrength = 150.0
 RoundsPerSecond = 1.0
