+++ scripts/weapons/Missile.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using MechJamIV;
 
-public partial class Missile : Grenade
+public partial class Missile : ExplosiveProjectile
 {
 
 	[Export]
@@ -12,8 +12,6 @@ public partial class Missile : Grenade
 
 	[Export]
 	public float TurnSpeed { get; set; } = 3_000f;
-
-	public override PickupType WeaponType { get => PickupType.Missile; }
 
 	#region Node references
